+++ Sources/SwifterSwift/SwiftStdlib/FloatExtensions.swift
@@ -52,15 +52,3 @@ func ** (lhs: Float, rhs: Float) -> Float {
     // http://nshipster.com/swift-operators/
     return pow(lhs, rhs)
 }
-
-// swiftlint:disable identifier_name
-prefix operator √
-/// SwifterSwift: Square root of float.
-///
-/// - Parameter float: float value to find square root for
-/// - Returns: square root of given float.
-public prefix func √ (float: Float) -> Float {
-    // http://nshipster.com/swift-operators/
-    return sqrt(float)
-}
-// swiftlint:enable identifier_name